compilation errors on 32bit
storage/innobase/fsp/fsp0fsp.cc:3108:58: error: conversion to ‘uint16_t {aka short unsigned int}’ from ‘unsigned int’ may alter its value [-Werror=conversion]
storage/innobase/handler/handler0alter.cc:10202:25: error: conversion to ‘unsigned char:7’ from ‘ulint {aka unsigned int}’ may alter its value [-Werror=conversion]

+++ storage/innobase/fsp/fsp0fsp.cc
@@ -3103,7 +3103,8 @@ fsp_truncate_list(fil_space_t *space, buf_block_t *header,
     {
       /* Remove the truncated free list extents */
       err= flst_remove(
-             header, FSP_HEADER_OFFSET + hdr_offset, descr_block,
+             header,
+             static_cast<uint16_t>(FSP_HEADER_OFFSET + hdr_offset), descr_block,
              static_cast<uint16_t>(descr - descr_block->page.frame
                                    + XDES_FLST_NODE), mtr);
       if (hdr_offset == FSP_FREE)
